Move testing the projector matrices to the appropriate test file.
