+++ tests/qadence/test_matrices.py
@@ -8,6 +8,7 @@ import strategies as st
 import torch
 from hypothesis import given, settings
 from metrics import ATOL_32, ATOL_E6
+from torch import Tensor
 
 from qadence import Parameter, QuantumCircuit, VariationalParameter, run
 from qadence.blocks import (
@@ -18,6 +19,8 @@ from qadence.blocks import (
     embedding,
 )
 from qadence.blocks.block_to_tensor import (
+    IMAT,
+    ZMAT,
     TensorType,
     _block_to_tensor_embedded,
     block_to_tensor,
@@ -39,6 +42,7 @@ from qadence.operations import (
     MCRX,
     MCRY,
     MCRZ,
+    MCZ,
     PHASE,
     RX,
     RY,
@@ -48,6 +52,7 @@ from qadence.operations import (
     HamEvo,
     I,
     N,
+    Projector,
     S,
     T,
     Toffoli,
@@ -65,6 +70,404 @@ def _calc_mat_vec_wavefunction(
 ) -> torch.Tensor:
     mat = block_to_tensor(block, values, tuple(range(n_qubits)))
     return torch.einsum("bij,kj->bi", mat, init_state)
+
+
+@pytest.mark.parametrize(
+    "projector, exp_projector_mat",
+    [
+        (
+            Projector(bra="1", ket="1", qubit_support=0),
+            torch.tensor([[0.0 + 0.0j, 0.0 + 0.0j], [0.0 + 0.0j, 1.0 + 0.0j]]),
+        ),
+        (
+            Projector(bra="10", ket="01", qubit_support=(2, 3)),
+            torch.tensor(
+                [
+                    [0.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j],
+                    [0.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j],
+                    [0.0 + 0.0j, 1.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j],
+                    [0.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j],
+                ]
+            ),
+        ),
+        (
+            N(0),
+            (IMAT - ZMAT) / 2.0,
+        ),
+        (
+            CNOT(0, 1),
+            torch.tensor(
+                [
+                    [
+                        [1.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j],
+                        [0.0 + 0.0j, 1.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j],
+                        [0.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j, 1.0 + 0.0j],
+                        [0.0 + 0.0j, 0.0 + 0.0j, 1.0 + 0.0j, 0.0 + 0.0j],
+                    ]
+                ]
+            ),
+        ),
+        (
+            MCZ((0, 1), 2),
+            torch.tensor(
+                [
+                    [
+                        [
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            -1.0 + 0.0j,
+                        ],
+                    ]
+                ]
+            ),
+        ),
+        (
+            MCRX((0, 1), 2, 1.0),
+            torch.tensor(
+                [
+                    [
+                        [
+                            1.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                        ],
+                        [
+                            0.0000 + 0.0000j,
+                            1.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                        ],
+                        [
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            1.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                        ],
+                        [
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            1.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                        ],
+                        [
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            1.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                        ],
+                        [
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            1.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                        ],
+                        [
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.8776 + 0.0000j,
+                            0.0000 - 0.4794j,
+                        ],
+                        [
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 + 0.0000j,
+                            0.0000 - 0.4794j,
+                            0.8776 + 0.0000j,
+                        ],
+                    ]
+                ]
+            ),
+        ),
+        (
+            CSWAP(0, 1, 2),
+            torch.tensor(
+                [
+                    [
+                        [
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                        ],
+                    ]
+                ]
+            ),
+        ),
+        (
+            Toffoli((0, 1), 2),
+            torch.tensor(
+                [
+                    [
+                        [
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                        ],
+                        [
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            0.0 + 0.0j,
+                            1.0 + 0.0j,
+                            0.0 + 0.0j,
+                        ],
+                    ]
+                ]
+            ),
+        ),
+    ],
+)
+def test_projector_tensor(projector: AbstractBlock, exp_projector_mat: Tensor) -> None:
+    projector_mat = block_to_tensor(projector)
+    assert torch.allclose(projector_mat, exp_projector_mat, atol=1.0e-4)
 
 
 @given(st.batched_digital_circuits())
