+++ tests/qadence/test_operators.py
@@ -18,7 +18,7 @@ from qadence.blocks import (
     kron,
     to_openfermion,
 )
-from qadence.blocks.block_to_tensor import IMAT, ZMAT, block_to_tensor
+from qadence.blocks.block_to_tensor import block_to_tensor
 from qadence.circuit import QuantumCircuit
 from qadence.operations import (
     CNOT,
@@ -26,12 +26,10 @@ from qadence.operations import (
     CRX,
     CRY,
     CRZ,
-    CSWAP,
     MCPHASE,
     MCRX,
     MCRY,
     MCRZ,
-    MCZ,
     RX,
     RY,
     RZ,
@@ -96,404 +94,6 @@ def crz_eigenvals(p: float, n_qubits: int = 2) -> torch.Tensor:
 
 def cphase_eigenvals(p: float, n_qubits: int = 2) -> torch.Tensor:
     return torch.cat((torch.ones(2**n_qubits - 1), eigenval(2.0 * p).conj()))
-
-
-@pytest.mark.parametrize(
-    "projector, exp_projector_mat",
-    [
-        (
-            Projector(bra="1", ket="1", qubit_support=0),
-            torch.tensor([[0.0 + 0.0j, 0.0 + 0.0j], [0.0 + 0.0j, 1.0 + 0.0j]]),
-        ),
-        (
-            Projector(bra="10", ket="01", qubit_support=(2, 3)),
-            torch.tensor(
-                [
-                    [0.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j],
-                    [0.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j],
-                    [0.0 + 0.0j, 1.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j],
-                    [0.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j],
-                ]
-            ),
-        ),
-        (
-            N(0),
-            (IMAT - ZMAT) / 2.0,
-        ),
-        (
-            CNOT(0, 1),
-            torch.tensor(
-                [
-                    [
-                        [1.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j],
-                        [0.0 + 0.0j, 1.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j],
-                        [0.0 + 0.0j, 0.0 + 0.0j, 0.0 + 0.0j, 1.0 + 0.0j],
-                        [0.0 + 0.0j, 0.0 + 0.0j, 1.0 + 0.0j, 0.0 + 0.0j],
-                    ]
-                ]
-            ),
-        ),
-        (
-            MCZ((0, 1), 2),
-            torch.tensor(
-                [
-                    [
-                        [
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            -1.0 + 0.0j,
-                        ],
-                    ]
-                ]
-            ),
-        ),
-        (
-            MCRX((0, 1), 2, 1.0),
-            torch.tensor(
-                [
-                    [
-                        [
-                            1.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                        ],
-                        [
-                            0.0000 + 0.0000j,
-                            1.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                        ],
-                        [
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            1.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                        ],
-                        [
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            1.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                        ],
-                        [
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            1.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                        ],
-                        [
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            1.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                        ],
-                        [
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.8776 + 0.0000j,
-                            0.0000 - 0.4794j,
-                        ],
-                        [
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 + 0.0000j,
-                            0.0000 - 0.4794j,
-                            0.8776 + 0.0000j,
-                        ],
-                    ]
-                ]
-            ),
-        ),
-        (
-            CSWAP(0, 1, 2),
-            torch.tensor(
-                [
-                    [
-                        [
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                        ],
-                    ]
-                ]
-            ),
-        ),
-        (
-            Toffoli((0, 1), 2),
-            torch.tensor(
-                [
-                    [
-                        [
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                        ],
-                        [
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            0.0 + 0.0j,
-                            1.0 + 0.0j,
-                            0.0 + 0.0j,
-                        ],
-                    ]
-                ]
-            ),
-        ),
-    ],
-)
-def test_projector_tensor(projector: AbstractBlock, exp_projector_mat: Tensor) -> None:
-    projector_mat = block_to_tensor(projector)
-    assert torch.allclose(projector_mat, exp_projector_mat, atol=1.0e-4)
 
 
 # PyQTorch only supports single qubit projectors.
